konfirmasi sweetalert

+++ src/components/view/Store.vue
@@ -10,6 +10,13 @@
               <div class="card">
                 <div class="card-header">
                   <h3 class="card-title">Data Store</h3>
+                  <router-link to="addstore" class="card-tools"> 
+                    <button
+                        class="btn btn-success"
+                    >
+                        Add Store
+                    </button>
+                </router-link>
                 </div>
                 <!-- /.card-header -->
                 <div class="card-body">
@@ -29,7 +36,7 @@
                         <td>{{ item.store.address }}</td>
                         <td>
                           <a href="#"><i class="fas fa-edit blue" style="padding-right:10px;"></i></a>
-                          <a href="#"  @click="deleteData(item.store.id)" class="text-danger">
+                          <a href="#" @click="deleteData(item.store.id)" class="text-danger">
                             <i class="fas fa-trash-alt red"></i></a>
                         </td>
                       </tr>
@@ -61,6 +68,8 @@
   import "datatables.net-dt/js/dataTables.dataTables"
   import "datatables.net-dt/css/jquery.dataTables.min.css"
   import $ from 'jquery';
+  import Swal from 'sweetalert2';
+
 
 
   export default {
@@ -91,38 +100,67 @@
           .then(({
             data
           }) => {
-          this.stores = data.data;
-           
-        })
+            this.stores = data.data;
+
+          })
           .catch((err) => {
             console.log(err)
           });
       },
       deleteData(id) {
-        axios.delete('https://api-kasirin.jaggs.id/api/stores/' + id)
-        .then(res => {
-          this.load()
-          console.log(res);
+        Swal.fire({
+          title: "Anda Yakin Ingin Menghapus Store Ini ?",
+          text: "Klik Batal untuk Membatalkan Penghapusan",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          cancelButtonText: "Batal",
+          confirmButtonText: "Hapus"
+        }).then(result => {
+          if (result.value) {
+            axios.delete('https://api-kasirin.jaggs.id/api/stores/' + id)
+              .then(res => {
+                Swal.fire(
+                  "Terhapus",
+                  "Store Anda Sudah Terhapus",
+                  "success"
+                );
+                this.load();
+                console.log(res);
+              })
+              .catch((err) => {
+                Swal.fire(
+                  "Gagal",
+                  "Store Anda Gagal Terhapus",
+                  "warning"
+                );
+                console.log(err)
+              })
+          } else {
+            Swal.fire(
+              "Gagal",
+              "Store Anda Gagal Terhapus",
+              "warning"
+            );
+          }
         })
-        .catch((err) => {
-        console.log(err)
-        });
       }
     },
 
     mounted() {
-      this.load( axios
-          .get("https://api-kasirin.jaggs.id/api/user-stores?user_id=" + localStorage.getItem('id'), {
-            headers: {
-              Authorization: 'Bearer ' + localStorage.getItem('access_token')
+      this.load(axios
+        .get("https://api-kasirin.jaggs.id/api/user-stores?user_id=" + localStorage.getItem('id'), {
+          headers: {
+            Authorization: 'Bearer ' + localStorage.getItem('access_token')
 
-            }
-          })
-          .then(({
-            data
-          }) => {
+          }
+        })
+        .then(({
+          data
+        }) => {
           this.stores = data.data;
-            $(function () {
+          $(function () {
             $('#datatable').DataTable({
               language: {
                 info: "",
@@ -136,9 +174,9 @@
             })
           })
         })
-          .catch((err) => {
-            console.log(err)
-          }));
+        .catch((err) => {
+          console.log(err)
+        }));
     },
   }
 </script>
